#21 [FEAT] 투표 생성 api 구현 완료

+++ src/main/java/hackaton/ip_backend/domain/Survey.java
@@ -50,7 +50,7 @@ public class Survey extends BaseEntity {
 	private String secondOption;
 
 	@Column(name = "ip_amount")
-	private int ipAmount;
+	private Long ipAmount;
 
 	@Column(name = "prize", length = 100)
 	private String prize;
